removed Testing from directory tree
