--- coastal/SFINCS/Testing/coastalforcing/config_full_cycle_1.yaml
@@ -1,40 +1,0 @@
-# -----------------------------------------------------------------------------
-# User Configuration File for Coastal Forcing
-#
-# This file controls model type, hydrology source, meteo source, 
-# coastal water level source, domain selection, and output paths.
-# -----------------------------------------------------------------------------
-coastal_model: sfincs                   # Options: sfincs (schism not implemented)
-meteo_source: nwm_retro                   # Options: nwm_retro, nwm_ana
-hydrology_source: nwm                   # Options: nwm, ngen
-coastal_water_level_source: tpxo       # Options: stofs, tpxo, glofs (great lakes only)
-domain_file: texas_example              # Options: see domain_lists/{coastal_model}/
-
-start_time: 2020-05-11T00-00-00Z        # Simulation start time in format YYYY-MM-DDTHH-MM-SSZ
-end_time: 2020-05-15T00-00-00Z          # Simulation end time in format YYYY-MM-DDTHH-MM-SSZ
-
-# These should be externally mounted
-raw_download_dir: ../data/raw/          # Directory for storing raw downloaded data
-sim_dir: ../run/                        # Directory for storing all processed model files ready to run
-
-
-# TPXO Info 
-tpxo_relative_path: process_data/TPXO
-tpxo_model_control: Model_tpxo10_atlas  # inside relative path w.r.t tpxo_relative_path
-ld_library_path: /contrib/software/gcc/8.5.0/lib64:/contrib/software/netcdf/4.7.4/lib
-
-#NOAA Info
-geojson_path: "schism_overlap_poly.geojson"
-utm_epsg: 32614
-interval_minutes: 6
-out_filename: "noaa.out"
-station_types:
-  - waterlevels
-  - currents
-plot_domain_facecolor: "lightblue"
-plot_figsize: [10, 8]
-plot_dpi: 150
-label_offset_deg: 0.01
-
-
-run_sfincs: true
