+++ coastal/SFINCS/coastalforcing/domain_lists/schism/all.yaml
@@ -1,0 +1,29 @@
+# This file is intended to be an example of how one could list domains available to run
+domain:
+  - name: atlgulf
+    path: ..\..\..\CoastalModels\schism\atlgulf\
+    region: atlantic_gulf   # General region or basin group
+    # coastal_points: <eventual link to hydrofabric?>
+    notes: >
+      SCHISM domain covering the full Atlantic and Gulf coastlines.
+
+  - name: pacific
+    path: ..\..\..\CoastalModels\schism\pacific\
+    region: pacific   # General region or basin group
+    # coastal_points: <eventual link to hydrofabric?>
+    notes: >
+      SCHISM domain covering the full US Pacific coastline.
+
+  - name: hawaii
+    path: ..\..\..\CoastalModels\schism\hawaii\
+    region: hawaii   # General region or basin group
+    # coastal_points: <eventual link to hydrofabric?>
+    notes: >
+      SCHISM domain covering the Hawaii coastline.
+
+  - name: prvi
+    path: ..\..\..\CoastalModels\schism\prvi\
+    region: prvi   # General region or basin group
+    # coastal_points: <eventual link to hydrofabric?>
+    notes: >
+      SCHISM domain covering the Puerto Rican and USVI coastlines.
