+++ coastal/SFINCS/coastalforcing/process_data/TPXO/setup.tpxo_sfincs
@@ -1,0 +1,10 @@
+process_data/TPXO/Model_tpxo10_atlas
+process_data/TPXO/tpxo_lat_lon_time
+z
+
+AP
+oce
+1
+process_data/TPXO/tpxo_out.txt
+
+
