+++ coastal/SFINCS/coastalmodels/sfincs/great_lakes_example_bndonly/sfincs.bnd
@@ -1,0 +1,16 @@
+368968.748	4586980.499
+372574.978	4584919.796
+377800.333	4584772.603
+382657.705	4587274.885
+382657.705	4589041.202
+392078.063	4589114.799
+396125.873	4590660.326
+399879.296	4593015.416
+404810.265	4594634.540
+409888.426	4596180.067
+415187.378	4597946.384
+420191.943	4596253.664
+423798.173	4595149.715
+429685.897	4594708.136
+436015.200	4594928.926
+438664.675	4595149.715
