+++ coastal/SFINCS/coastalmodels/sfincs/texas_example/sfincs.bnd
@@ -1,0 +1,15 @@
+    752497.0   3134390.2
+	758243.4   3135771.4
+    765751.2   3140335.2
+    773259.1   3144899.1
+    780766.9   3149462.9
+    788274.8   3154026.8
+    795782.7   3158590.7
+    803290.5   3163154.5
+    810798.4   3167718.4
+    818306.2   3172282.2
+    825814.1   3176846.1
+    833321.9   3181409.9
+    840829.8   3185973.8
+    848337.6   3190537.6
+    849816.2   3194578.7
