+++ coastal/SFINCS/coastalmodels/sfincs/texas_example/sfincs.inp
@@ -1,0 +1,54 @@
+latitude             = 0.0
+qtrfile              = sfincs.nc
+tref                 = 20250601 000000
+tstart               = 20250601 000000
+tstop                = 20250608 000000
+tspinup              = 3600.0
+dtmapout             = 3600.0
+dthisout             = 600.0
+dtrstout             = 0.0
+dtmaxout             = 0.0
+trstout              = -999.0
+dtwnd                = 1800.0
+alpha                = 0.5
+theta                = 1.0
+viscosity            = 0
+nuvisc               = 0.01
+huthresh             = 0.01
+zsini                = 0.0
+qinf                 = 0.0
+igperiod             = 120.0
+rhoa                 = 1.25
+rhow                 = 1024.0
+dtmax                = 999.0
+maxlev               = 999.0
+bndtype              = 1
+advection            = 0
+baro                 = 0
+pavbnd               = 101200.0
+gapres               = 101200.0
+advlim               = 9999.9
+stopdepth            = 1000.0
+crsgeo               = 0
+epsg                 = 32614
+wiggle_suppression   = 0
+storevel             = 1
+storemeteo           = 0
+nonh                 = 0
+sbgfile              = sfincs.sbg
+bndfile              = sfincs.bnd
+bzsfile              = sfincs.bzs
+obsfile              = sfincs.obs
+srcfile              = sfincs.src
+disfile              = sfincs.dis
+amufile              = sfincs.amu
+amvfile              = sfincs.amv
+amprfile             = sfincs.ampr
+ampfile              = sfincs.amp
+snapwave_gammaig     = 1.0
+snapwave_gammax      = 1.0
+inputformat          = bin
+outputformat         = net
+cdnrb                = 3
+cdwnd                = 0.0 28.0 50.0 
+cdval                = 0.001 0.0025 0.0025
