+++ coastal/SFINCS/coastalmodels/sfincs/texas_example/sfincs.obs
@@ -1,0 +1,7 @@
+830344.95 3187383.41 "Sargent (8772985)"
+801500.14 3179759.77 "Matagorda City (8773146)"
+733642.95 3170505.01 "Port Lavaca (8773259)"
+755758.90 3150019.47 "Port O'Connor (8773701)"
+761521.61 3147395.72 "Matagorda Bay Entrance Channel (8773767)"
+758914.37 3149643.75 "Matagorda Ship Channel Marker 19 (mg0101)"
+761531.97 3147370.44 "Matagorda Entrance Channel (mg0201)"
