+++ coastal/SFINCS/coastalmodels/sfincs/texas_example/sfincs_ngen.src
@@ -1,0 +1,56 @@
+805561.884	3197062.245	"nex-2414735 poi-86738"
+808288.319	3205910.032	"nex-2414675 poi-87012"
+805848.293	3211177.105	"nex-2414620 poi-86803"
+808770.331	3212158.407	"nex-2414564 poi-89094"
+810600.091	3214374.755	"nex-2414609 poi-87373"
+810763.966	3215479.62	"nex-2414608 poi-87363"
+808991.343	3202784.827	"nex-2414690 poi-86737"
+717564.043	3170882.348	"nex-2438337 poi-10690"
+714558.229	3178519.009	"nex-2438356 poi-10671"
+717737.722	3179586.389	"nex-2438357 poi-23888"
+717737.722	3179586.389	"nex-2438357 poi-49078"
+715575.852	3195159.156	"nex-2438396 poi-48946"
+717156.581	3195268.677	"nex-2438397 poi-19686"
+719044.477	3192970.268	"nex-2438398 poi-19660"
+718762.964	3192088.804	"nex-2438399 poi-10668"
+721908.031	3188071.772	"nex-2438405 poi-19693"
+723040.334	3191549.44	"nex-2438449 poi-19633"
+723040.334	3191549.44	"nex-2438449 poi-19620"
+740348.526	3192334.141	"nex-2438534 poi-8064"
+721145.218	3197276.052	"nex-2438447 poi-48920"
+721145.218	3197276.052	"nex-2438447 poi-10666"
+729313.855	3197870.849	"nex-2437754 poi-87344"
+738374.978	3198300.524	"nex-2437804 poi-87352"
+736497.844	3198921.087	"nex-2437798 poi-87764"
+774683.493	3199447.883	"nex-2438570 poi-10664"
+775450.249	3201786.099	"nex-2438569 poi-26331"
+775450.249	3201786.099	"nex-2438569 poi-10618"
+776746.181	3206515.246	"nex-2438568 poi-10626"
+777628.277	3199130.249	"nex-2438586 poi-23537"
+774607.641	3205747.113	"nex-2438598 poi-7861"
+782455.287	3197148.131	"nex-2438634 poi-10591"
+772584.172	3194950.877	"nex-2438639 poi-10575"
+761313.868	3197971.097	"nex-2438691 poi-10574"
+764807.678	3198297.082	"nex-2438676 poi-10599"
+763511.62	3199810.661	"nex-2438675 poi-10595"
+747584.602	3199685.774	"nex-2438715 poi-9993"
+763185.493	3200428.469	"nex-2438674 poi-10442"
+749140.645	3198249.043	"nex-2438716 poi-9992"
+754688.737	3196549.695	"nex-2438727 poi-9997"
+789249.029	3199184.484	"nex-2430682 poi-43458"
+804155.491	3200753.202	"nex-2414729 poi-86735"
+802777.747	3202205.91	"nex-2414731 poi-40728"
+797443.591	3202407.609	"nex-2414721 poi-86834"
+791501.824	3205718.042	"nex-2430691 poi-89089"
+783984.652	3200627.794	"nex-2438627 poi-8159"
+784268.11	3199305.81	"nex-2438628 poi-10578"
+739796.655	3202683.306	"nex-2437797 poi-89100"
+725002.282	3204255.546	"nex-2437752 poi-19699"
+741068.688	3205015.455	"nex-2437796 poi-87354"
+740533.919	3205205.492	"nex-2437795 poi-87773"
+725451.13	3205570.304	"nex-2437751 poi-59280"
+725451.13	3205570.304	"nex-2437751 poi-87371"
+724561.654	3206604.333	"nex-2437750 poi-19702"
+724813.61	3209002.524	"nex-2437749 poi-87315"
+738290.681	3209916.581	"nex-2437811 poi-87355"
+738455.144	3210595.377	"nex-2437793 poi-87347"
