+++ coastal/SFINCS/coastalmodels/sfincs/texas_example/sfincs_nwm.src
@@ -1,0 +1,48 @@
+807142.08 3204919.62 "1614588"
+807818.85 3203166.89 "1614624"
+803376.13 3202998.24 "1614632"
+807685.56 3201964.20 "1614660"
+805403.42 3200859.41 "1614692"
+808626.38 3201211.11 "1614710"
+790715.42 3205227.92 "3765398"
+792317.82 3196481.30 "3765432"
+741446.59 3202249.66 "7847403"
+739740.83 3198652.48 "7847427"
+721605.71 3192632.33 "9349323"
+719110.39 3179772.07 "9350001"
+776497.69 3201991.28 "9355372"
+777889.91 3200763.07 "9355456"
+766615.12 3196656.49 "9355544"
+807427.80 3206380.76 "1614576"
+796643.06 3202988.70 "1614622"
+796864.43 3201997.05 "1614648"
+737751.82 3201164.07 "7847415"
+722930.22 3188476.35 "9349389"
+774202.61 3201936.96 "9355388"
+757615.48 3197042.74 "9355524"
+809817.51 3213673.25 "1612376"
+790583.00 3206021.05 "3765394"
+783916.84 3199922.60 "9355446"
+756955.20 3197396.12 "9355512"
+763451.57 3195941.10 "9355622"
+739572.63 3190625.42 "9355778"
+798941.02 3204350.73 "1614604"
+716628.52 3181518.43 "9349989"
+814899.17 3217621.37 "1612350"
+733781.20 3198414.69 "7842859"
+775300.14 3204870.95 "9355300"
+807492.49 3210427.96 "1614480"
+800172.95 3205427.92 "1614584"
+762221.88 3201154.27 "9355398"
+747712.25 3202126.08 "9355422"
+715951.41 3179400.47 "9349449"
+808608.56 3212220.60 "1612394"
+752344.20 3200342.52 "9355410"
+740683.00 3210012.35 "7847373"
+716204.26 3195233.28 "9349291"
+776626.55 3204815.52 "9355310"
+724909.29 3207830.04 "7842817"
+721487.98 3195748.36 "9349289"
+738545.36 3210581.22 "7846485"
+790963.42 3206464.02 "3765766"
+790117.02 3201606.93 "3765414"
