+++ coastal/SFINCS/requirements.txt
@@ -1,0 +1,37 @@
+attrs==25.3.0
+certifi==2025.8.3
+cffi==1.17.1
+cfgrib==0.9.15.0
+cftime==1.6.4.post1
+charset-normalizer==3.4.3
+click==8.2.1
+colorama==0.4.6
+contourpy==1.3.3
+cycler==0.12.1
+eccodes==2.43.0
+findlibs==0.1.2
+fonttools==4.59.0
+geopandas==1.1.1
+h5netcdf==1.6.4
+h5py==3.14.0
+idna==3.10
+kiwisolver==1.4.9
+matplotlib==3.10.5
+netCDF4==1.7.2
+numpy==2.3.2
+packaging==25.0
+pandas==2.3.1
+pillow==11.3.0
+pycparser==2.22
+pyogrio==0.11.1
+pyparsing==3.2.3
+pyproj==3.7.1
+python-dateutil==2.9.0.post0
+pytz==2025.2
+PyYAML==6.0.2
+requests==2.32.4
+shapely==2.1.1
+six==1.17.0
+tzdata==2025.2
+urllib3==2.5.0
+xarray==2025.7.1
